# server/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: SkillHub API
  description: |
    SkillHub backend HTTP API.  Covers native portal routes (/api/v1/),
    web aliases (/api/web/), CLI routes (/api/cli/v1/), tool-facing
    routes for the miqro CLI (/api/tool/v1/), ClawHub compatibility,
    well-known discovery endpoints, and frontend page-oriented read
    models (/api/v1/frontend/*).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/cli/v1/auth/whoami:
    get:
      summary: CLI whoami — returns authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: CLI-compatible user principal
  /api/cli/v1/skills/search:
    get:
      summary: CLI search with installability filter
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: SearchResult

  # --- Tool API (miqro CLI protocol surface) ---
...
  /api/v1/skills/{skillID}/ci/runs:
    get:
      tags: [AgentCI]
      summary: List CI pipeline runs for a skill
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: size, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Paginated pipeline run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunListResult"
  /api/v1/skills/{skillID}/ci/runs/{runID}:
    get:
      tags: [AgentCI]
      summary: Get a single pipeline run by ID
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: runID, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Pipeline run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRun"
  /api/v1/skills/{skillID}/ci/runs/{runID}/checks:
    get:
      tags: [AgentCI]
      summary: List check runs for a pipeline run
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: runID, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Check run list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckRun"
  /api/v1/skills/{skillID}/ci/checks/{checkID}:
    get:
      tags: [AgentCI]
      summary: Get a single check run by ID
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: checkID, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Check run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckRun"
  /api/v1/skills/{skillID}/ci/checks/{checkID}/artifacts:
    get:
      tags: [AgentCI]
      summary: List artifacts for a check run
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: checkID, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Artifact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckArtifact"
  /api/v1/skills/{skillID}/ci/gates:
    get:
      tags: [AgentCI]
      summary: Evaluate CI gates for a skill
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - { name: skillID, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: trigger, in: query, schema: { type: string } }
        - { name: versionId, in: query, schema: { type: integer, format: int64 } }
        - { name: releaseId, in: query, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Gate evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateEvalResult"

  # ── Release publish with gates ─────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API token (format: sk_...)"
    sessionCookie:
      type: apiKey
      in: cookie
      name: skillhub_session
  schemas:
...
    PipelineRun:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pipelineId: { type: integer, format: int64 }
        skillId: { type: integer, format: int64 }
        versionId: { type: integer, format: int64 }
        releaseId: { type: integer, format: int64 }
        triggerType: { type: string }
        triggeredBy: { type: string }
        status: { type: string }
        checkCount: { type: integer }
        passedCount: { type: integer }
        failedCount: { type: integer }
        skippedCount: { type: integer }
        startedAt: { type: string }
        completedAt: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    PipelineRunListResult:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/PipelineRun"
        totalCount: { type: integer, format: int64 }
        page: { type: integer }
        size: { type: integer }
    CheckRun:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pipelineRunId: { type: integer, format: int64 }
        skillId: { type: integer, format: int64 }
        versionId: { type: integer, format: int64 }
        releaseId: { type: integer, format: int64 }
        name: { type: string }
        runnerType: { type: string }
        status: { type: string }
        conclusion: { type: string }
        summary: { type: string }
        isBlocking: { type: boolean }
        startedAt: { type: string }
        completedAt: { type: string }
        durationMs: { type: integer, format: int64 }
        createdAt: { type: string }
        updatedAt: { type: string }
    CheckArtifact:
      type: object
      properties:
        id: { type: integer, format: int64 }
        checkRunId: { type: integer, format: int64 }
        name: { type: string }
        contentType: { type: string }
        size: { type: integer, format: int64 }
        storageKey: { type: string }
        createdAt: { type: string }
    GatePolicyResult:
      type: object
      properties:
        policyId: { type: integer, format: int64 }
        policyName: { type: string }
        passed: { type: boolean }
        reason: { type: string }
    GateEvalResult:
      type: object
      properties:
        passed: { type: boolean }
        reason: { type: string }
        policyResults:
          type: array
          items:
            $ref: "#/components/schemas/GatePolicyResult"
...